+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/parameter/impl/ShardingGeneratedKeyInsertValueParameterRewriter.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.sharding.rewrite.parameter.impl;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import lombok.Setter;
-import org.apache.shardingsphere.sharding.route.ShardingRouteContext;
+import org.apache.shardingsphere.sharding.route.engine.context.ShardingRouteContext;
 import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementContext;
 import org.apache.shardingsphere.sql.parser.relation.statement.impl.InsertSQLStatementContext;
 import org.apache.shardingsphere.sharding.rewrite.aware.ShardingRouteContextAware;
